pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "setuptools-scm"]

[project]
authors = [
  {name = "Matt Kramer", email = "mkramer@anaconda.com"}
]
dependencies = [
  "cogapp<=3.4.1",
  "typer<0.13"
]
description = "Python pre-commit hooks from Anaconda"
license = {text = "Apache-2.0"}
name = "anaconda_pre_commit_hooks"
readme = "README.md"
requires-python = ">=3.8"
version = "0.1.0"

[project.scripts]
generate-renovate-annotations = "anaconda_pre_commit_hooks.add_renovate_annotations:app"
run-cog = "anaconda_pre_commit_hooks.run_cog:main"

[tool.mypy]
disallow_untyped_defs = true
files = [
  "src/**/*.py"
]
python_version = "3.9"

[tool.pytest.ini_options]
addopts = [
  "--cov",
  "--color=yes",
  "--cov-report=xml:./coverage.xml",
  "--cov-report=term-missing"
]
filterwarnings = ["error"]
norecursedirs = ["env"]
pythonpath = "src/"
xfail_strict = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
extend-select = ["I"]
